requirements.txt
requests==2.19.1
Pillow==5.3.0
